requirements.txt
cohere==4.34
python-dotenv==0.21.0
weaviate-client
qdrant-client
